# server/controllers/CMO_db/generated/OrganizationControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE OrganizationControllerGenerated.js PLEASE EDIT ../OrganizationController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import OrganizationModel from "../../../models/CMO_db/OrganizationModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import OrganizationController from "../OrganizationController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/organization`;
    router.post(baseUrl + "", authorize([]), OrganizationController.create);
    router.delete(baseUrl + "/:id", authorize([]), OrganizationController.delete);
    router.get(baseUrl + "/findByCountry/:key", authorize([]), OrganizationController.findByCountry);
    router.get(baseUrl + "/findByLessor/:key", authorize([]), OrganizationController.findByLessor);
    router.get(baseUrl + "/findByPartner/:key", authorize([]), OrganizationController.findByPartner);
    router.get(baseUrl + "/findBycountry/:key", authorize([]), OrganizationController.findBycountry);
    router.get(baseUrl + "/findBylessor/:key", authorize([]), OrganizationController.findBylessor);
    router.get(baseUrl + "/findBypartner/:key", authorize([]), OrganizationController.findBypartner);
    router.get(baseUrl + "/:id", authorize([]), OrganizationController.get);
    router.get(baseUrl + "", authorize([]), OrganizationController.list);
    router.post(baseUrl + "/:id", authorize([]), OrganizationController.update);
  },


  // CRUD METHODS


  /**
  * OrganizationModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await OrganizationModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await OrganizationModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.findByCountry
  *   @description CRUD ACTION findByCountry
  *   @param Objectid key Id of model to search for
  *
  */
  findByCountry: async (req, res) => {
    try {
      const result = await OrganizationModel.findByCountry(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.findByLessor
  *   @description CRUD ACTION findByLessor
  *   @param Objectid key Id of model to search for
  *
  */
  findByLessor: async (req, res) => {
    try {
      const result = await OrganizationModel.findByLessor(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.findByPartner
  *   @description CRUD ACTION findByPartner
  *   @param Objectid key Id of model to search for
  *
  */
  findByPartner: async (req, res) => {
    try {
      const result = await OrganizationModel.findByPartner(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.findBycountry
  *   @description CRUD ACTION findBycountry
  *   @param Objectid key Id of model to search for
  *
  */
  findBycountry: async (req, res) => {
    try {
      const result = await OrganizationModel.findBycountry(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.findBylessor
  *   @description CRUD ACTION findBylessor
  *   @param Objectid key Id of model to search for
  *
  */
  findBylessor: async (req, res) => {
    try {
      const result = await OrganizationModel.findBylessor(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.findBypartner
  *   @description CRUD ACTION findBypartner
  *   @param Objectid key Id of model to search for
  *
  */
  findBypartner: async (req, res) => {
    try {
      const result = await OrganizationModel.findBypartner(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await OrganizationModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * OrganizationModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await OrganizationModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * OrganizationModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await OrganizationModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;